requirements.txt
streamlit
pandas
geopandas
pydeck
fpdf2
qrcode